refactor(app): split app.go into domain files
app.go (1293 lines) → 6 domain files:
app_tasks.go, app_agents.go, app_projects.go,
app_orchestrator.go, app_planning.go, app_reviews.go.
Methods stay on *App receiver — no logic changes.
