+++ app_tasks.go
@@ -1,0 +1,79 @@
+package main
+
+import (
+	"slices"
+
+	"github.com/Automaat/synapse/internal/audit"
+	"github.com/Automaat/synapse/internal/task"
+)
+
+func (a *App) ListTasks() ([]task.Task, error) {
+	return a.tasks.List()
+}
+
+func (a *App) GetTask(id string) (task.Task, error) {
+	return a.tasks.Get(id)
+}
+
+func (a *App) CreateTask(title, body, mode string) (task.Task, error) {
+	t, err := a.tasks.Create(title, body, mode)
+	if err != nil {
+		return t, err
+	}
+	a.logAudit(audit.EventTaskCreated, t.ID, "", map[string]any{"title": title, "mode": mode})
+	if t.Status == task.StatusTodo {
+		a.logger.Info("auto-triage.start", "task_id", t.ID, "title", t.Title)
+		a.wg.Go(func() {
+			if triageErr := a.TriageTask(t.ID); triageErr != nil {
+				a.logger.Error("auto-triage.failed", "task_id", t.ID, "err", triageErr)
+			}
+		})
+	}
+	return t, nil
+}
+
+func (a *App) UpdateTask(id string, updates map[string]any) (task.Task, error) {
+	var prevStatus string
+	if newStatus, ok := updates["status"].(string); ok {
+		if prev, getErr := a.tasks.Get(id); getErr == nil {
+			prevStatus = string(prev.Status)
+			if prevStatus != newStatus {
+				a.logAudit(audit.EventTaskStatusChanged, id, "", map[string]any{"from": prevStatus, "to": newStatus})
+			}
+		}
+	}
+	t, err := a.tasks.Update(id, updates)
+	if err != nil {
+		return t, err
+	}
+	if t.Status == task.StatusPlanning {
+		a.logger.Info("auto-plan.start", "task_id", t.ID, "title", t.Title)
+		a.wg.Go(func() {
+			if planErr := a.PlanTask(t.ID); planErr != nil {
+				a.logger.Error("auto-plan.failed", "task_id", t.ID, "err", planErr)
+			}
+		})
+	}
+	if t.Status == task.StatusInProgress && !a.agents.HasRunningAgentForTask(t.ID) && !slices.Contains(t.Tags, "review") {
+		a.logger.Info("auto-implement.start", "task_id", t.ID, "title", t.Title)
+		a.wg.Go(func() {
+			if _, err := a.StartAgent(t.ID, t.AgentMode, "Implement this task. When done, create a draft PR with `gh pr create --draft`."); err != nil {
+				a.logger.Error("auto-implement.failed", "task_id", t.ID, "err", err)
+			}
+		})
+	}
+	if t.Status == task.StatusDone {
+		a.wg.Go(func() { a.cleanupWorktree(t.ID) })
+	}
+	return t, nil
+}
+
+func (a *App) DeleteTask(id string) error {
+	a.logger.Info("task.delete", "task_id", id)
+	a.logAudit(audit.EventTaskDeleted, id, "", nil)
+	if err := a.tasks.Delete(id); err != nil {
+		a.logger.Error("task.delete.failed", "task_id", id, "err", err)
+		return err
+	}
+	return nil
+}
